requirements.txt
fastapi
uvicorn
python-multipart
img2pdf
Pillow
reportlab